// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  ORGANIZER: 'ORGANIZER',
  USER: 'USER',
  MODERATOR: 'MODERATOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EventVisibility = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE'
} as const

export type EventVisibility = (typeof EventVisibility)[keyof typeof EventVisibility]


export const ParticipationStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  BANNED: 'BANNED'
} as const

export type ParticipationStatus = (typeof ParticipationStatus)[keyof typeof ParticipationStatus]


export const PaymentStatus = {
  UNPAID: 'UNPAID',
  PAID: 'PAID',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const InvitationStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  DECLINED: 'DECLINED'
} as const

export type InvitationStatus = (typeof InvitationStatus)[keyof typeof InvitationStatus]


export const Currency = {
  USD: 'USD',
  BDT: 'BDT',
  AED: 'AED',
  EUR: 'EUR',
  GBP: 'GBP'
} as const

export type Currency = (typeof Currency)[keyof typeof Currency]